Confidence intervals for Reporting Odds Ratio

Mainly for use in ror. Produces (symmetric, normality based) confidence bounds for the ROR, for a passed probability. Called twice in ror to create confidence intervals.

Usage

ci_for_ror(a, b, c, d, conf_lvl_probs)

Arguments

a

Number of reports for the specific drug and event (i.e. the observed count).

b

Number of reports with the drug, without the event

c

Number of reports without the drug, with the event

d

Number of reports without the drug, without the event

conf_lvl_probs

The probabilities of the normal distribution, based on a passed confidence level (conf_lvl) in ror. If sgn_lvl = .95 in ror, quantiles of the normal distribution will be extracted at sgn_lvl_probs of 0.025 and 0.975.

Value

The credibility interval specified by input parameters.

See also